+++ NEWS.md
@@ -8,13 +8,14 @@ output:
 News for Package 'SDLfilter'
 =========
 
-#### Version 1.2.1.9008 (2020-02-10)
+#### Version 1.2.1.9009 (2020-02-24)
 
 * added four new functions to assess sample sizes of animal tracking data (boot_overlap, boot_area, asymptote, percent_vol).
-* added a new example data "curtis"
+* added a new example data "curtis".
 * added a new argument "type" in the depthfilter function.
 * renamed some functions (ddfilter.loop -> ddfilter_loop; ddfilter.speed -> ddfilter_speed; dupfilter.exact -> dupfilter_exact; dupfilter.qi -> dupfilter_qi; dupfilter.space -> dupfilter_space; dupfilter.time -> dupfilter_time, distantfilter -> distfilter).
 * renamed an argument in ddfilter (maxvlp -> vmaxlp).
+* improved the plotMap function.
 * improved the R documentations.
 
 #### Version 1.2.1 (2019-02-19)